requirements.txt
opencv-python==4.1.2.30
scipy==1.1.0
packaging==20.1
tqdm==4.42.1
PyYAML==5.3
xmltodict==0.12.0
pandas==1.0.1
mat4py==0.5.0